+++ CHANGES.md
@@ -22,6 +22,7 @@ Bugfix 🐛:
  - Initial sync is not retried correctly when there is some network error. (#2632)
  - Fix switch theme issue, and white field issue (#2599, #2528)
  - Fix request too large Uri error when joining a room
+ - Render markdown in room list (#452)
 
 Translations 🗣:
  -